# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Context-Aware PQC Risk Assessment API
  version: 0.1.0
  license:
    name: Internal Capstone Project
    url: https://github.com/ehdgks0627/crypto-scanner
  description: |
    API contract for the PQC Risk Assessment System.

    `docs/api/openapi.yaml` is the design source of truth. The Django Ninja
    implementation must expose an equivalent `/api/openapi.json`.
servers:
  - url: /api
security:
  - ApiToken: []
...
paths:
...
  /discoveries:
    get:
      tags: [Discovery]
      summary: List discovery jobs
      operationId: listDiscoveries
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Discovery list
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryPage'
        '400':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Discovery]
      summary: Start CIDR discovery
      operationId: createDiscovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryCreate'
            examples:
              default:
                externalValue: ./examples/discovery-create.json
      responses:
        '202':
          description: Discovery job accepted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelope'
              examples:
                accepted:
                  externalValue: ./examples/discovery-job-accepted.json
        '400':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/Error'
  /discoveries/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Discovery]
      summary: Get discovery status
      operationId: getDiscovery
      responses:
        '200':
          description: Discovery status
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Discovery'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        default:
          $ref: '#/components/responses/Error'
  /discoveries/{id}/endpoints:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Discovery]
      summary: List endpoints found by discovery
      operationId: listDiscoveredEndpoints
      parameters:
        - name: promoted
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Discovered endpoints
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveredEndpointPage'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /discoveries/{id}/promote:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Discovery]
      summary: Promote discovered endpoints to targets
      operationId: promoteDiscoveredEndpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryPromoteRequest'
      responses:
        '201':
          description: Promotion result
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryPromoteResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
...
  /agents/register:
    post:
      tags: [Agents]
      summary: Register an agent
      operationId: registerAgent
      security:
        - BootstrapToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegister'
            examples:
              default:
                externalValue: ./examples/agent-register.json
      responses:
        '201':
          description: Agent registered for a new hostname. Returns a raw token once.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegisterResult'
              examples:
                created:
                  externalValue: ./examples/agent-register-result-created.json
        '200':
          description: Existing hostname re-registered. Agent metadata is updated and token is rotated.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegisterResult'
              examples:
                token_rotated:
                  externalValue: ./examples/agent-register-result-rotated.json
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '422':
          $ref: '#/components/responses/Unprocessable'
        default:
          $ref: '#/components/responses/Error'
  /agents:
    get:
      tags: [Agents]
      summary: List agents
      operationId: listAgents
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Agent list
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPage'
              examples:
                default:
                  externalValue: ./examples/agents-list.json
        '400':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /agents/{id}:
    parameters:
      - $ref: '#/components/parameters/Uuid'
    get:
      tags: [Agents]
      summary: Get agent details
      operationId: getAgent
      responses:
        '200':
          description: Agent details
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Agents]
      summary: Deactivate an agent
      operationId: deleteAgent
      responses:
        '204':
          description: Deactivated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /agents/{id}/heartbeat:
    parameters:
      - $ref: '#/components/parameters/Uuid'
    post:
      tags: [Agents]
      summary: Record agent heartbeat
      operationId: heartbeatAgent
      security:
        - AgentBearer: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentHeartbeat'
            examples:
              default:
                externalValue: ./examples/agent-heartbeat.json
      responses:
        '200':
          description: Heartbeat accepted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentHeartbeatResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
    BootstrapToken:
      type: apiKey
      in: header
      name: X-Bootstrap-Token
    AgentBearer:
      type: http
      scheme: bearer
  headers:
    XRequestId:
      description: Request trace id. If the client sends `X-Request-Id`, the server echoes it.
      schema:
        type: string
        format: uuid
    NoStore:
      description: Polling responses are not cacheable.
      schema:
        type: string
      example: no-store
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Uuid:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: Field name, prefixed with `-` for descending order.
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              externalValue: ./examples/error-validation.json
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InvalidToken:
      description: Invalid or missing token
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unprocessable:
      description: Request is semantically invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Required backend dependency is unavailable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ProtocolHint:
      type: string
      enum: [TLS, SSH, IKE, SMTP, IMAP, POP3, UNKNOWN]
    DetectedProtocol:
      type: string
      enum: [HTTPS, TLS, SSH, IKE, SMTP, SMTPS, IMAP, IMAPS, POP3, POP3S, MQTT, PostgreSQL, UNKNOWN]
...
    Sensitivity:
      type: string
      enum: [low, medium, high, critical]
    Criticality:
      type: string
      enum: [low, medium, high, critical]
    Exposure:
      type: string
      enum: [public_internet, dmz, internal_network, air_gapped]
    JobStatus:
      type: string
      enum: [PENDING, RUNNING, COMPLETED, FAILED, CANCELLED]
    JobKind:
      type: string
      enum: [scan_job, discovery, recompute]
...
    ErrorResponse:
      type: object
      required: [error, message, details]
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
        details:
          type: object
          description: Machine-readable error details. Empty object when there are no details.
          additionalProperties: true
    TargetContext:
      type: object
      properties:
        sensitivity:
          type: string
          allOf:
            - $ref: '#/components/schemas/Sensitivity'
          nullable: true
        lifespan_years:
          type: integer
          minimum: 0
          nullable: true
        criticality:
          type: string
          allOf:
            - $ref: '#/components/schemas/Criticality'
          nullable: true
        exposure:
          type: string
          allOf:
            - $ref: '#/components/schemas/Exposure'
          nullable: true
        service_role:
          type: string
          nullable: true
...
    PageBase:
      type: object
      required: [total, offset, limit]
      properties:
        total:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
    DiscoveryCreate:
      type: object
      required: [cidr]
      properties:
        cidr:
          type: string
          example: 172.20.0.0/24
        ports:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 65535
        include_default_ports:
          type: boolean
          description: Defaults to true when omitted.
    Discovery:
      type: object
      required: [id, job_id, cidr, port_list, status, created_at, started_at]
      properties:
        id:
          type: integer
        job_id:
          type: integer
          description: API-visible AsyncJob id used for polling or cancellation.
        cidr:
          type: string
        port_list:
          type: array
          items:
            type: integer
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobProgress'
          nullable: true
        created_at:
          type: string
          format: date-time
          description: Discovery request creation time.
        started_at:
          type: string
          format: date-time
          nullable: true
          description: Worker execution start time. Null while status is PENDING.
        finished_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
    DiscoveryPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Discovery'
    DiscoveredEndpoint:
      type: object
      required:
        - id
        - ip
        - port
        - banner_metadata
        - promoted
      properties:
        id:
          type: integer
        ip:
          type: string
          format: ip
        port:
          type: integer
          minimum: 1
          maximum: 65535
        detected_protocol:
          type: string
          allOf:
            - $ref: '#/components/schemas/DetectedProtocol'
          nullable: true
        banner_metadata:
          type: object
          additionalProperties: true
        promoted:
          type: boolean
        target_id:
          type: integer
          nullable: true
        suggested_protocol_hint:
          type: string
          allOf:
            - $ref: '#/components/schemas/ProtocolHint'
          nullable: true
        suggested_host:
          type: string
          nullable: true
    DiscoveredEndpointPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/DiscoveredEndpoint'
    DiscoveryPromotion:
      type: object
      required: [endpoint_id, host, protocol_hint]
      properties:
        endpoint_id:
          type: integer
        host:
          type: string
        protocol_hint:
          $ref: '#/components/schemas/ProtocolHint'
        context:
          $ref: '#/components/schemas/TargetContext'
        agent_enabled:
          type: boolean
          description: Defaults to false when omitted.
    DiscoveryPromoteRequest:
      type: object
      required: [promotions]
      properties:
        promotions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/DiscoveryPromotion'
    DiscoveryPromoteResult:
      type: object
      required: [promoted, skipped]
      properties:
        promoted:
          type: array
          items:
            type: object
            required: [endpoint_id, target_id]
            properties:
              endpoint_id:
                type: integer
              target_id:
                type: integer
        skipped:
          type: array
          items:
            type: object
            required: [endpoint_id, reason]
            properties:
              endpoint_id:
                type: integer
              reason:
                type: string
    JobProgress:
      type: object
      properties:
        completed:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        current_target:
          type: string
          nullable: true
        current_scanner:
          type: string
          nullable: true
      additionalProperties: true
    JobResource:
      type: object
      description: Resource created by the job. For recompute, the AsyncJob itself is the resource and id equals JobEnvelope.id.
      required: [kind, id]
      properties:
        kind:
          $ref: '#/components/schemas/JobKind'
        id:
          type: integer
          description: Resource id for the job kind. For recompute this equals the API-visible job id.
    JobResult:
      type: object
      description: Completed-job result only. PENDING and RUNNING jobs must return null for JobEnvelope.result.
      properties:
        snapshot_id:
          type: integer
          nullable: true
        discovery_id:
          type: integer
          nullable: true
        updated_scores_count:
          type: integer
          description: Number of RiskScore rows updated by a completed recompute job.
          nullable: true
      additionalProperties: false
    JobEnvelope:
      type: object
      required: [id, kind, resource, status, progress, started_at, cancel_requested_at, finished_at, result, error]
      properties:
        id:
          type: integer
          description: API-visible asynchronous job id, unique across all job kinds.
        kind:
          $ref: '#/components/schemas/JobKind'
        resource:
          $ref: '#/components/schemas/JobResource'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobProgress'
          nullable: true
          description: Always present. Null while PENDING or when no progress can be reported.
        started_at:
          type: string
          format: date-time
          nullable: true
        cancel_requested_at:
          type: string
          format: date-time
          nullable: true
          description: Set when cancellation has been requested but a running worker has not finished cleanup yet.
        finished_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobResult'
          nullable: true
          example:
            snapshot_id: 56
        error:
          type: string
          nullable: true
...
    AgentRegister:
      type: object
      required: [hostname, capabilities]
      description: Uses X-Bootstrap-Token. New hostnames create an agent; existing hostnames update metadata and rotate the agent token.
      properties:
        hostname:
          type: string
          maxLength: 253
        agent_url:
          type: string
          format: uri
          nullable: true
        capabilities:
          type: array
          items:
            type: string
        os_distribution:
          type: string
          nullable: true
    AgentRegisterResult:
      type: object
      required: [id, agent_token, registration_action, token_rotated_at]
      properties:
        id:
          type: string
          format: uuid
        agent_token:
          type: string
          description: Raw token returned once at registration time.
        registration_action:
          type: string
          enum: [created, token_rotated]
        token_rotated_at:
          type: string
          format: date-time
    Agent:
      type: object
      required:
        - id
        - hostname
        - agent_url
        - capabilities
        - registered_at
        - token_rotated_at
        - last_seen
        - active
        - is_stale
      properties:
        id:
          type: string
          format: uuid
        hostname:
          type: string
        agent_url:
          type: string
          format: uri
          nullable: true
        capabilities:
          type: array
          items:
            type: string
        os_distribution:
          type: string
          nullable: true
        registered_at:
          type: string
          format: date-time
        token_rotated_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        active:
          type: boolean
        is_stale:
          type: boolean
    AgentPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Agent'
    AgentHeartbeat:
      type: object
      properties:
        uptime_sec:
          type: integer
          minimum: 0
        version:
          type: string
      additionalProperties: true
    AgentHeartbeatResult:
      type: object
      required: [received_at]
      properties:
        received_at:
          type: string
          format: date-time
...